+++ src/apphosting/config.spec.ts
@@ -3,11 +3,13 @@ import * as sinon from "sinon";
 import * as yaml from "yaml";
 import * as path from "path";
 import * as fsImport from "../fsutils";
+import * as csmImport from "../gcp/secretManager";
 import * as promptImport from "../prompt";
 import * as dialogs from "./secrets/dialogs";
 import * as config from "./config";
 import { NodeType } from "yaml/dist/nodes/Node";
-import { AppHostingYamlConfig } from "./yaml";
+import { AppHostingYamlConfig, toEnvList } from "./yaml";
+import { FirebaseError } from "../error";
 
 describe("config", () => {
   describe("discoverBackendRoot", () => {
@@ -240,60 +242,226 @@ env:
       ]);
     });
   });
-  describe("loadConfigForEnvironment", () => {
-    let loadFromFileStub: sinon.SinonStub;
-    let baseAppHostingYaml: AppHostingYamlConfig;
-    let stagingAppHostingYaml: AppHostingYamlConfig;
-    let baseEnv: Record<string, Omit<config.Env, "variable">>;
-    let stagingEnv: Record<string, Omit<config.Env, "variable">>;
+
+  describe("maybeGenerateEmulatorsYaml", () => {
+    let discoverBackendRoot: sinon.SinonStub;
+    let overrideChosenEnv: sinon.SinonStub;
+    let loadFromFile: sinon.SinonStub;
+    let store: sinon.SinonStub;
+    let fs: sinon.SinonStubbedInstance<typeof fsImport>;
+    let prompt: sinon.SinonStubbedInstance<typeof promptImport>;
+
+    const existingYaml = AppHostingYamlConfig.empty();
+    existingYaml.env = {
+      VAR: { value: "value" },
+      API_KEY: { secret: "api-key" },
+      API_KEY2: { secret: "api-key2" },
+    };
 
     beforeEach(() => {
-      baseEnv = {
-        ENV_1: { value: "base_env_1" },
-        ENV_3: { value: "base_env_3" },
-        SECRET_1: { secret: "base_secret_1" },
-        SECRET_2: { secret: "base_secret_2" },
-        SECRET_3: { secret: "base_secret_3" },
-      };
-      baseAppHostingYaml = AppHostingYamlConfig.empty();
-      baseAppHostingYaml.env = { ...baseEnv };
-
-      stagingEnv = {
-        ENV_1: { value: "staging_env_1" },
-        ENV_2: { value: "staging_env_2" },
-        SECRET_1: { secret: "staging_secret_1" },
-        SECRET_2: { secret: "staging_secret_2" },
-      };
-      stagingAppHostingYaml = AppHostingYamlConfig.empty();
-      stagingAppHostingYaml.env = { ...stagingEnv };
-
-      loadFromFileStub = sinon.stub(AppHostingYamlConfig, "loadFromFile");
-      loadFromFileStub.callsFake(async (filePath) => {
-        if (filePath?.includes("apphosting.staging.yaml")) {
-          return Promise.resolve(stagingAppHostingYaml);
-        }
-        return Promise.resolve(baseAppHostingYaml);
-      });
+      discoverBackendRoot = sinon.stub(config, "discoverBackendRoot");
+      overrideChosenEnv = sinon.stub(config, "overrideChosenEnv");
+      store = sinon.stub(config, "store");
+      loadFromFile = sinon.stub(AppHostingYamlConfig, "loadFromFile");
+      fs = sinon.stub(fsImport);
+      prompt = sinon.stub(promptImport);
     });
 
     afterEach(() => {
       sinon.verifyAndRestore();
     });
 
-    it("returns a config that complies with the expected precendence", async () => {
-      const resultingConfig = await config.loadConfigForEnvironment(
-        "/parent/cwd/apphosting.staging.yaml",
-        "/parent/cwd/apphosting.yaml",
-      );
-      expect(resultingConfig.env).to.deep.equal({
-        ...baseEnv,
-        ...stagingEnv,
-      });
+    it("noops if emulators.yaml already exists", async () => {
+      discoverBackendRoot.withArgs("/project").returns("/project");
+      fs.fileExistsSync.withArgs(`/project/${config.APPHOSTING_EMULATORS_YAML_FILE}`).returns(true);
+
+      await config.maybeGenerateEmulatorYaml("projectId", "/project");
+
+      expect(prompt.confirm).to.not.have.been.called;
+      expect(store).to.not.have.been.called;
     });
 
-    it("returns appropriate config if only base file was selected", async () => {
-      const resultingConfig = await config.loadConfigForEnvironment("/parent/cwd/apphosting.yaml");
-      expect(resultingConfig.env).to.deep.equal(baseEnv);
+    // This allows us to prompt to give devs access to prod keys
+    it("returns existing config even if the user does not create apphosting.emulator.yaml", async () => {
+      discoverBackendRoot.withArgs("/project").returns("/project");
+      fs.fileExistsSync
+        .withArgs(`/project/${config.APPHOSTING_EMULATORS_YAML_FILE}`)
+        .returns(false);
+      // Do not create emulator file
+      prompt.confirm.resolves(false);
+      loadFromFile.resolves(existingYaml);
+
+      await expect(
+        config.maybeGenerateEmulatorYaml("projectId", "/project"),
+      ).to.eventually.deep.equal(toEnvList(existingYaml.env));
+    });
+
+    it("returns overwritten config", async () => {
+      discoverBackendRoot.withArgs("/project").returns("/project");
+      fs.fileExistsSync
+        .withArgs(`/project/${config.APPHOSTING_EMULATORS_YAML_FILE}`)
+        .returns(false);
+      loadFromFile.resolves(existingYaml);
+      // Create emulator file
+      prompt.confirm.resolves(true);
+      overrideChosenEnv.resolves({
+        API_KEY2: { secret: "test-api-key2" },
+      });
+      store.resolves();
+
+      await expect(
+        config.maybeGenerateEmulatorYaml("projectId", "/project"),
+      ).to.eventually.deep.equal([
+        { variable: "VAR", value: "value" },
+        { variable: "API_KEY", secret: "api-key" },
+        { variable: "API_KEY2", secret: "test-api-key2" },
+      ]);
+
+      expect(overrideChosenEnv.firstCall.args[1]).to.deep.equal({
+        VAR: { value: "value" },
+        API_KEY: { secret: "api-key" },
+        API_KEY2: { secret: "api-key2" },
+      });
+      expect(store).to.have.been.called;
+      const emulatorYaml = store.firstCall.args[1] as yaml.Document;
+      expect(emulatorYaml.toJSON()).to.deep.equal({
+        env: [{ variable: "API_KEY2", secret: "test-api-key2" }],
+      });
+    });
+  });
+
+  describe("overrideChosenEnv", () => {
+    let csm: sinon.SinonStubbedInstance<typeof csmImport>;
+    let prompt: sinon.SinonStubbedInstance<typeof promptImport>;
+
+    beforeEach(() => {
+      csm = sinon.stub(csmImport);
+      prompt = sinon.stub(promptImport);
+    });
+
+    afterEach(() => {
+      sinon.verifyAndRestore();
+    });
+
+    it("noops with no envs", async () => {
+      await expect(config.overrideChosenEnv(undefined, {})).to.eventually.deep.equal({});
+
+      expect(promptImport.promptOnce).to.not.have.been.called;
+      expect(csmImport.getSecret).to.not.have.been.called;
+    });
+
+    it("noops with no selected envs", async () => {
+      const originalEnv: Record<string, Omit<config.Env, "variable">> = {
+        VARIABLE: { value: "value" },
+        API_KEY: { secret: "api-key" },
+      };
+
+      prompt.promptOnce.onFirstCall().resolves([]);
+
+      await expect(config.overrideChosenEnv(undefined, originalEnv)).to.eventually.deep.equal({});
+
+      expect(prompt.promptOnce).to.have.been.calledOnce;
+      expect(csm.secretExists).to.not.have.been.called;
+    });
+
+    it("can override plaintext values", async () => {
+      const originalEnv: Record<string, config.Env> = {
+        VARIABLE: { variable: "VARIABLE", value: "value" },
+        VARIABLE2: { variable: "VARIABLE2", value: "value2" },
+      };
+
+      prompt.promptOnce.onFirstCall().resolves(["VARIABLE2"]);
+      prompt.promptOnce.onSecondCall().resolves("new-value2");
+
+      await expect(config.overrideChosenEnv(undefined, originalEnv)).to.eventually.deep.equal({
+        VARIABLE2: { variable: "VARIABLE2", value: "new-value2" },
+      });
+
+      expect(prompt.promptOnce).to.have.been.calledTwice;
+      expect(csmImport.secretExists).to.not.have.been.called;
+    });
+
+    it("throws when trying to overwrite secrets without knowing the project", async () => {
+      const originalEnv: Record<string, config.Env> = {
+        API_KEY: { variable: "API_KEY", secret: "api-key" },
+      };
+
+      prompt.promptOnce.onFirstCall().resolves(["API_KEY"]);
+
+      await expect(config.overrideChosenEnv(undefined, originalEnv)).to.be.rejectedWith(
+        FirebaseError,
+        /Need a project ID to overwrite a secret./,
+      );
+    });
+
+    it("can create new secrets", async () => {
+      const originalEnv: Record<string, config.Env> = {
+        API_KEY: { variable: "API_KEY", secret: "api-key" },
+      };
+
+      prompt.promptOnce.onFirstCall().resolves(["API_KEY"]);
+      prompt.promptOnce.onSecondCall().resolves("test-api-key");
+      csm.secretExists.withArgs("project", "test-api-key").resolves(false);
+      prompt.promptOnce.onThirdCall().resolves("plaintext secret value");
+
+      await expect(config.overrideChosenEnv("project", originalEnv)).to.eventually.deep.equal({
+        API_KEY: { variable: "API_KEY", secret: "test-api-key" },
+      });
+
+      expect(prompt.promptOnce).to.have.been.calledThrice;
+      expect(csm.secretExists).to.have.been.calledOnce;
+      expect(csm.createSecret).to.have.been.calledOnce;
+      expect(csm.addVersion).to.have.been.calledOnce;
+      expect(csm.addVersion.getCall(0).args[2]).to.equal("plaintext secret value");
+    });
+
+    it("can create new secrets after warning about reuse", async () => {
+      const originalEnv: Record<string, config.Env> = {
+        API_KEY: { variable: "API_KEY", secret: "api-key" },
+      };
+
+      prompt.promptOnce.onCall(0).resolves(["API_KEY"]);
+      prompt.promptOnce.onCall(1).resolves("test-api-key");
+      csm.secretExists.withArgs("project", "test-api-key").resolves(true);
+      prompt.promptOnce.onCall(2).resolves("pick-new");
+      prompt.promptOnce.onCall(3).resolves("test-api-key2");
+      prompt.promptOnce.onCall(4).resolves("plaintext secret value");
+
+      await expect(config.overrideChosenEnv("project", originalEnv)).to.eventually.deep.equal({
+        API_KEY: { variable: "API_KEY", secret: "test-api-key2" },
+      });
+
+      expect(prompt.promptOnce.callCount).to.equal(5);
+      expect(csm.secretExists).to.have.been.calledTwice;
+      expect(csm.createSecret).to.have.been.calledOnce;
+      expect(csm.addVersion).to.have.been.calledOnce;
+      expect(csm.addVersion.getCall(0).args[2]).to.equal("plaintext secret value");
+    });
+
+    it("can reuse secrets", async () => {
+      const originalEnv: Record<string, config.Env> = {
+        API_KEY: { variable: "API_KEY", secret: "api-key" },
+      };
+
+      prompt.promptOnce.onFirstCall().resolves(["API_KEY"]);
+      prompt.promptOnce.onSecondCall().resolves("test-api-key");
+      csm.secretExists.withArgs("project", "test-api-key").resolves(true);
+      prompt.promptOnce.onThirdCall().resolves("reuse");
+
+      await expect(config.overrideChosenEnv("project", originalEnv)).to.eventually.deep.equal({
+        API_KEY: { variable: "API_KEY", secret: "test-api-key" },
+      });
+
+      expect(prompt.promptOnce).to.have.been.calledThrice;
+      expect(csm.secretExists).to.have.been.calledOnce;
+      expect(csm.createSecret).to.not.have.been.called;
+      expect(csm.addVersion).to.not.have.been.called;
+    });
+
+    it("suggests test key names", () => {
+      expect(config.suggestedTestKeyName("GOOGLE_GENAI_API_KEY")).to.equal(
+        "test-google-genai-api-key",
+      );
     });
   });
 });